Infers ethnicity using DNA methylation microarray data

pl_infer_ethnicity Uses a glmnet model to predict ethnicity using DNA methylation data.

pl_infer_ethnicity(betas, threshold = 0.75)

Arguments

betas
n x m dataframe of methylation values on the beta scale (0, 1), where the variables are arranged in rows, and samples in columns. Should contain all 1860 predictors and be normalized with NOOB and BMIQ.

threshold
A probability threshold ranging from (0, 1) to call samples 'ambiguous'. Defaults to 0.75.

Value

a [tibble][tibble::tibble-package]

Details

Predicts self-reported ethnicity from 3 classes: Africans, Asians, and Caucasians, using placental DNA methylation data measured on the Infinium 450k/EPIC methylation array. Will return membership probabilities that often reflect genetic ancestry composition.

The input data should contain all 1860 predictors (cpgs) of the final GLMNET model.

It's recommended to use the same normalization methods used on the training data: NOOB and BMIQ.

Examples

## To predict ethnicity on 450k/850k samples

# Load placenta DNAm data
data(pl_betas)
pl_infer_ethnicity(pl_betas)

#> [1] "1860 of 1860 predictors present."

#> Loading required package: Matrix

#> # A tibble: 24 x 7
#>    Sample_ID Predicted_ethni~ Predicted_ethni~ Prob_African Prob_Asian
#>    <chr>     <chr>            <chr>                   <dbl>      <dbl>
#>  1 GSM19449~ Caucasian        Caucasian            0.00331    0.0164  
#>  2 GSM19449~ Caucasian        Caucasian            0.000772   0.000514
#>  3 GSM19449~ Caucasian        Caucasian            0.000806   0.000699
#>  4 GSM19449~ Caucasian        Caucasian            0.000883   0.000792
#>  5 GSM19449~ Caucasian        Caucasian            0.000885   0.00130 
#>  6 GSM19449~ Caucasian        Caucasian            0.000852   0.000973
#>  7 GSM19449~ Caucasian        Caucasian            0.000902   0.00176 
#>  8 GSM19449~ Caucasian        Caucasian            0.00174    0.00223 
#>  9 GSM19449~ Caucasian        Caucasian            0.000962   0.00231 
#> 10 GSM19449~ Caucasian        Caucasian            0.00287    0.00356 
#> # ... with 14 more rows, and 2 more variables: Prob_Caucasian <dbl>,
#> #   Highest_Prob <dbl>