predictPreeclampsia

Uses 45 CpGs to predict early preeclampsia (PE delivered before or at 34 weeks of gestation) on placental DNA methylation microarray data.

predictPreeclampsia(betas, ...)

Arguments

betas

matrix or array of methylation values on the beta scale (0, 1), where the variables are arranged in rows, and samples in columns.

...

feeds into outersect function

Value

produces a list with components detailed in the mixOmics::predict R documentation

Details

Assigns the class labels "early-PE" or "normotensive" to each sample and returns a class probability.

It is recommended that users apply beta-mixture quantile normalization (BMIQ) to their data

prior to prediction. This was the normalization method used on the training data.

Examples

# To predict early preeclampsia on 450k/850k samples

# Load data
library(ExperimentHub)
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     combine, intersect, setdiff, union
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, saveRDS, setdiff, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: AnnotationHub
#> Loading required package: BiocFileCache
#> Loading required package: dbplyr
#> 
#> Attaching package: ‘dbplyr’
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     ident, sql
eh <- ExperimentHub()
query(eh, "eoPredData")
#> ExperimentHub with 3 records
#> # snapshotDate(): 2024-10-24
#> # $dataprovider: University of British Columbia
#> # $species: Homo sapiens
#> # $rdataclass: mixo_splsda, matrix, data.frame
#> # additional mcols(): taxonomyid, genome, description,
#> #   coordinate_1_based, maintainer, rdatadateadded, preparerclass, tags,
#> #   rdatapath, sourceurl, sourcetype 
#> # retrieve records with, e.g., 'object[["EH8090"]]' 
#> 
#>            title      
#>   EH8090 | eoPredModel
#>   EH8403 | valBMIQ    
#>   EH8404 | valMeta    

# test object
x_test <- eh[['EH8403']]
#> eoPredData not installed.
#>   Full functionality, documentation, and loading of data might not be possible without installing
#> downloading 1 resources
#> retrieving 1 resource
#> loading from cache
x_test %>% predictPreeclampsia()
#> eoPredData not installed.
#>   Full functionality, documentation, and loading of data might not be possible without installing
#> downloading 1 resources
#> retrieving 1 resource
#> loading from cache
#> 45 of 45 predictive CpGs present.
#> BMIQ normalization is recommended for best results. If choosing other method, it is recommended to compare results to predictions on BMIQ normalized data.
#> # A tibble: 49 × 4
#>    Sample_ID   EOPE `Non-PE Preterm` PE_Status   
#>    <chr>      <dbl>            <dbl> <chr>       
#>  1 GSM2589533 0.670            0.330 EOPE        
#>  2 GSM2589535 0.768            0.232 EOPE        
#>  3 GSM2589536 0.807            0.193 EOPE        
#>  4 GSM2589538 0.784            0.216 EOPE        
#>  5 GSM2589540 0.386            0.614 Normotensive
#>  6 GSM2589541 0.649            0.351 EOPE        
#>  7 GSM2589545 0.653            0.347 EOPE        
#>  8 GSM2589546 0.348            0.652 Normotensive
#>  9 GSM2589549 0.797            0.203 EOPE        
#> 10 GSM2589552 0.485            0.515 Normotensive
#> # ℹ 39 more rows